requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.4
psycopg[binary]>=3.1
python-dotenv>=1.0
demucs>=4.0
gunicorn>=22.0
